move mini-cart into modal folder
