+++ web/app/containers/header/container.jsx
@@ -9,7 +9,7 @@ import throttle from 'lodash.throttle'
 import classnames from 'classnames'
 
 import * as headerActions from './actions'
-import * as miniCartActions from '../mini-cart/actions'
+import * as miniCartActions from '../../modals/mini-cart/actions'
 import {openModal} from 'progressive-web-sdk/dist/store/modals/actions'
 import {NAVIGATION_MODAL} from '../../modals/constants'
 import * as selectors from './selectors'
