+++ web/app/modals/mini-cart/_mini-cart.scss
