+++ web/app/modals/mini-cart/partials/mini-cart-header.jsx
