+++ web/app/modals/mini-cart/partials/mini-cart-product-list.jsx
