+++ web/app/styles/_templates.scss
@@ -8,7 +8,7 @@
 @import '../containers/home/home';
 @import '../containers/product-list/product-list';
 @import '../containers/header/header';
-@import '../containers/mini-cart/mini-cart';
+@import '../modals/mini-cart/mini-cart';
 @import '../containers/footer/footer';
 @import '../containers/navigation/navigation';
 @import '../containers/login/login';
